bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Adilson Pavan <adilson.pavan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.13.0"
python-telegram-bot = "^12.7"
sqlalchemy = "^1.3.17"
mysqlclient = "^1.4.6"


[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
